pyproject.toml
[project]
name = "elpm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "imbalanced-learn>=0.13.0",
    "ipykernel>=6.29.5",
    "mlxtend>=0.23.4",
    "numpy>=2.3.0",
    "pandas>=2.3.0",
    "plotly>=6.1.2",
    "pyarrow>=20.0.0",
    "python-dotenv>=1.1.0",
    "requests>=2.32.4",
    "scikit-learn>=1.6.1",
    "scipy>=1.15.3",
    "seaborn>=0.13.2",
    "statsmodels>=0.14.4",
    "streamlit>=1.46.0",
    "streamlit-option-menu>=0.4.0",
    "vadersentiment>=3.3.2",
]
